Moved UserModel from UserRepo to a new dedicated file: "db-models/user-model.ts"
