+++ src/db-models/user-model.ts
@@ -1,0 +1,14 @@
+import { Maybe } from '../types/graphql';
+
+export type UserModel = {
+	id: string;
+	name: string;
+	dob: string;
+	address?: Maybe<string>;
+	description?: Maybe<string>;
+	imageUrl?: Maybe<string>;
+	createdAt: string;
+	updatedAt: string;
+};
+
+export default UserModel;
